Move lookup handlers into a subdirectory for clarity
